requirements.txt
requests>=2.26.0,<3.0.0
rich==13.7.1
packaging>=23.0
PyYAML>=6.0.1